+++ src/erlydtl_runtime.erl
@@ -20,7 +20,7 @@
 
 find_value(Key, Data, Options) when is_atom(Key), is_tuple(Data) ->
     Rec = element(1, Data),
-    Info = proplists:get_value(record_info, Options),
+    Info = proplists:get_value(record_info, Options, []),
     case proplists:get_value(Rec, Info) of
         Fields when is_list(Fields), length(Fields) == size(Data) - 1 ->
             case proplists:get_value(Key, Fields) of
@@ -107,6 +107,32 @@ find_value(Key, Tuple) when is_tuple(Tuple) ->
                 _ ->
                     undefined
             end
+    end;
+find_value(Key, Map) ->
+    case erlang:is_builtin(erlang, is_map, 1) andalso erlang:is_map(Map) of
+        true  -> find_map_value(Key, Map);
+        false -> undefined
+    end.
+
+find_map_value(Key, Map) when is_atom(Key) ->
+    case maps:find(Key, Map) of
+        error           -> find_map_value(atom_to_list(Key), Map);
+        {ok, Value}     -> Value
+    end;
+find_map_value(Key, Map) when is_list(Key) ->
+    case maps:find(Key, Map) of
+        error           -> find_map_value(list_to_binary(Key), Map);
+        {ok, Value}     -> Value
+    end;
+find_map_value(Key, Map) when is_binary(Key) ->
+    case maps:find(Key, Map) of
+        error           -> undefined;
+        {ok, Value}     -> Value
+    end;
+find_map_value(Key, Map) ->
+    case maps:find(Key, Map) of
+        error           -> undefined;
+        {ok, Value}     -> Value
     end.
 
 fetch_value(Key, Data, Options) ->
@@ -118,28 +144,33 @@ fetch_value(Key, Data, Options, Default) ->
         Val -> Val
     end.
 
-find_deep_value([Key|Rest],Item) ->
-    case find_value(Key,Item) of
+find_deep_value(Key, Data) ->
+    find_deep_value(Key, Data, []).
+
+find_deep_value([Key|Rest], Item, Opts) ->
+    case find_value(Key, Item, Opts) of
         undefined -> undefined;
-        NewItem -> find_deep_value(Rest,NewItem)
+        NewItem -> find_deep_value(Rest, NewItem, Opts)
     end;
-find_deep_value([],Item) -> Item.
+find_deep_value([], Item, _Opts) -> Item.
 
 regroup(List, Attribute) ->
-    regroup(List, Attribute, []).
+    do_regroup(List, Attribute, [], []).
 
-regroup([], _, []) ->
-    [];
-regroup([], _, [[{grouper, LastGrouper}, {list, LastList}]|Acc]) ->
+regroup(List, Attribute, Options) ->
+    do_regroup(List, Attribute, Options, []).
+
+do_regroup([], _, _, []) -> [];
+do_regroup([], _, _, [[{grouper, LastGrouper}, {list, LastList}]|Acc]) ->
     lists:reverse([[{grouper, LastGrouper}, {list, lists:reverse(LastList)}]|Acc]);
-regroup([Item|Rest], Attribute, []) ->
-    regroup(Rest, Attribute, [[{grouper, find_deep_value(Attribute, Item)}, {list, [Item]}]]);
-regroup([Item|Rest], Attribute, [[{grouper, PrevGrouper}, {list, PrevList}]|Acc]) ->
-    case find_deep_value(Attribute, Item) of
+do_regroup([Item|Rest], Attribute, Options, []) ->
+    do_regroup(Rest, Attribute, Options, [[{grouper, find_deep_value(Attribute, Item, Options)}, {list, [Item]}]]);
+do_regroup([Item|Rest], Attribute, Options, [[{grouper, PrevGrouper}, {list, PrevList}]|Acc]) ->
+    case find_deep_value(Attribute, Item, Options) of
         Value when Value =:= PrevGrouper ->
-            regroup(Rest, Attribute, [[{grouper, PrevGrouper}, {list, [Item|PrevList]}]|Acc]);
+            do_regroup(Rest, Attribute, Options, [[{grouper, PrevGrouper}, {list, [Item|PrevList]}]|Acc]);
         Value ->
-            regroup(Rest, Attribute, [[{grouper, Value}, {list, [Item]}], [{grouper, PrevGrouper}, {list, lists:reverse(PrevList)}]|Acc])
+            do_regroup(Rest, Attribute, Options, [[{grouper, Value}, {list, [Item]}], [{grouper, PrevGrouper}, {list, lists:reverse(PrevList)}]|Acc])
     end.
 
 -spec init_translation(init_translation()) -> none | translate_fun().
@@ -243,6 +274,7 @@ are_equal(_, _) ->
 is_false("") -> true;
 is_false(false) -> true;
 is_false(undefined) -> true;
+is_false(null) -> true;
 is_false(0) -> true;
 is_false("0") -> true;
 is_false(<<"0">>) -> true;
@@ -361,12 +393,17 @@ increment_counter_stats([{counter, Counter}, {counter0, Counter0}, {revcounter, 
      {first, false}, {last, RevCounter0 =:= 1},
      {parentloop, Parent}].
 
-forloop(_Fun, [], _Parent) -> empty;
-forloop(Fun, Values, Parent) ->
+forloop(_Fun, [], _Parent, Default) -> Default;
+forloop(Fun, Values, Parent, _Default) ->
     push_ifchanged_context(),
-    Result = lists:mapfoldl(Fun, init_counter_stats(Values, Parent), Values),
+    {Result, _Acc} = lists:mapfoldl(Fun, init_counter_stats(Values, Parent), Values),
     pop_ifchanged_context(),
     Result.
+
+%% keep old version for backwards compatibility..
+forloop(_Fun, [], _Parent) -> empty;
+forloop(Fun, Values, Parent) ->
+    {forloop(Fun, Values, Parent, undefined), undefined}.
 
 push_ifchanged_context() ->
     IfChangedContextStack = case get(?IFCHANGED_CONTEXT_VARIABLE) of