+++ src/erlydtl_scanner.erl
@@ -36,7 +36,7 @@
 %%%-------------------------------------------------------------------
 -module(erlydtl_scanner).
 
-%% This file was generated 2014-04-15 19:15:09 UTC by slex 0.2.1.
+%% This file was generated 2014-12-16 18:46:16 UTC by slex 0.2.1-2-g7814678.
 %% http://github.com/erlydtl/slex
 -slex_source(["src/erlydtl_scanner.slex"]).
 
@@ -92,6 +92,7 @@ is_keyword(any, "context") -> true;
 is_keyword(any, "noop") -> true;
 is_keyword(close, "only") -> true;
 is_keyword(close, "parsed") -> true;
+is_keyword(close, "silent") -> true;
 is_keyword(close, "reversed") -> true;
 is_keyword(close, "openblock") -> true;
 is_keyword(close, "closeblock") -> true;
@@ -350,7 +351,7 @@ scan(" " ++ T, S, {R, C} = P,
      {in_verbatim_code, E} = St) ->
     {Tag, Backtrack} = E,
     scan(T, S, {R, C + 1},
-	 {in_verbatim_code, {Tag, [$  | Backtrack]}});
+	 {in_verbatim_code, {Tag, [$\s | Backtrack]}});
 scan("endverbatim%}" ++ T, S, {R, C} = P,
      {in_verbatim_code, E} = St)
     when element(1, E) =:= undefined ->
@@ -366,7 +367,8 @@ scan(" " ++ T, S, {R, C} = P,
     when element(3, E) =:= "" ->
     {Tag, Backtrack, EndTag} = E,
     scan(T, S, {R, C + 1},
-	 {in_endverbatim_code, {Tag, [$  | Backtrack], EndTag}});
+	 {in_endverbatim_code,
+	  {Tag, [$\s | Backtrack], EndTag}});
 scan([H | T], S, {R, C} = P,
      {in_endverbatim_code, E} = St)
     when H >= $a andalso H =< $z orelse
@@ -380,7 +382,7 @@ scan(" " ++ T, S, {R, C} = P,
     when element(1, E) =:= element(3, E) ->
     {Tag, Backtrack, Tag} = E,
     scan(T, S, {R, C + 1},
-	 {in_endverbatim_code, {Tag, [$  | Backtrack], Tag}});
+	 {in_endverbatim_code, {Tag, [$\s | Backtrack], Tag}});
 scan("%}" ++ T, S, {R, C} = P,
      {in_endverbatim_code, E} = St)
     when element(1, E) =:= element(3, E) ->
@@ -563,6 +565,14 @@ post_process([{open_tag, _, _} | _],
 post_process([{open_tag, _, _} | _],
 	     {identifier, _, L} = T, _) ->
     is_keyword(open_tag, T);
+post_process([{open_var, _, _} | _],
+	     {identifier, _, L} = T, _) ->
+    setelement(3, T,
+	       begin L1 = lists:reverse(L), L2 = to_atom(L1), L2 end);
+post_process([{'.', _} | _], {identifier, _, L} = T,
+	     _) ->
+    setelement(3, T,
+	       begin L1 = lists:reverse(L), L2 = to_atom(L1), L2 end);
 post_process(_, {identifier, _, L} = T, close_tag) ->
     is_keyword(close_tag, T);
 post_process(_, {identifier, _, L} = T, _) ->