+++ NEWS.md
@@ -5,6 +5,25 @@ suggested by the [GNU Coding
 Standards](http://www.gnu.org/prep/standards/html_node/NEWS-File.html#NEWS-File).
 
 
+## 0.10.0 (2014-12-20)
+
+* Fix issue with generated code for `for` loops (#167).
+
+* Fix issue with using keywords as attributes (#177), or as variables (#194).
+
+* Fix issue when including multiple templates extending a common base template (#176).
+
+* New `w` option for toggling compile time warnings. Currently, there is only one, `non_block_tag`,
+  which is triggered on any non-block data in an extends-template.
+
+* Add missing features to the `cycle` tag (#195) (still missing
+  independtly stepping the cycle value).
+
+* Support records in regroup tag (#191).
+
+* Support for maps (#196).
+
+
 ## 0.9.4 (2014-04-15)
 
 * Fix compile time variables and constants (#61)